Found at least one package whose jarjar treatment should differ between server and client versions. Here I am introducing a client jarjarrules copy with is identical to server but without org.apache.** zapped. This package is part of the Android runtime, so in the server it should be zapped, but in the client there is no replacement in the standard JRE.
Extra notes for the curious...
I considered creating three rules files: one contains the rules common to both
client and server, and two more files for the unique rules. But this would add
more steps and complication to the build process, as well as the code exporting
to svn.

Not zapping these files in the server jar changes the jar size from about 2.1 to
2.5MB

Affected files appear in the public repository as:
./android/jarjarrules_client
./android/jarjarrules_server

R=dxu,tkaizu


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1938